小程序-定时任务

+++ src/main/java/com/plat/paygate/shop/domain/PgSettle.java
@@ -29,7 +29,7 @@ public class PgSettle {
 
     private Date settleDate;
 
-    private Byte settleState;
+    private Integer settleState;
 
     private Date createTime;
